backend/main_config/wsgi.py
# main_config/wsgi.py

import os
from django.core.wsgi import get_wsgi_application

os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'main_config.settings')

application = get_wsgi_application()
